several new feature for Texture  - added version of load_from_memory that loads all layers and levels at once  - both load_from_memory functions now take a length argument, purely for error checking  - added method to load all data from an lz4 compressed binary file    - data is already in correct format so loading is MUCH faster for hdr images  - added method to save all data to an lz4 compressed binary file    - this is mainly for use by the sts editor for generating environment maps  - added getters for the internal vulkan image and sampler  - 2D texture json files are now required to specify size  - array and cube texture json files are now next to their image directory rather than inside add basic support for displaying images to imgui backend improve imgui shader to use just a vec2 for projection rather than a full mat4 move e5bgr9 compression function out of Texture.cpp into Colours.hpp changed ResourceCache::try_acquire to catch exceptions, so that it can print the key StagingBuffer can now be used a a destination (as well as a source) added RenderPass, Framebuffer and DescriptorSet create/allocate helpers to VulkanContext

+++ source/sqee/app/DebugOverlay.cpp
@@ -95,9 +95,9 @@ void DebugOverlay::show_imgui_widgets()
         {
             const auto& display = mFrameSubTimers[mFrameSubTimers.size() - lineNum - 1u].display;
             const char* strEnd = display.data() + display.length();
-            drawList->AddText(ImVec2(15, ImPlus::FromScreenBottom(95 + lineNum * 16)), IM_COL32_BLACK, display.data(), strEnd);
-            drawList->AddText(ImVec2(16, ImPlus::FromScreenBottom(96 + lineNum * 16)), IM_COL32_BLACK, display.data(), strEnd);
-            drawList->AddText(ImVec2(16, ImPlus::FromScreenBottom(96 + lineNum * 16)), IM_COL32_WHITE, display.data(), strEnd);
+            drawList->AddText(ImVec2(15, ImPlus::FromScreenBottom(float(95 + lineNum * 16))), IM_COL32_BLACK, display.data(), strEnd);
+            drawList->AddText(ImVec2(16, ImPlus::FromScreenBottom(float(96 + lineNum * 16))), IM_COL32_BLACK, display.data(), strEnd);
+            drawList->AddText(ImVec2(16, ImPlus::FromScreenBottom(float(96 + lineNum * 16))), IM_COL32_WHITE, display.data(), strEnd);
         }
 
         const char* strEnd = mFrameSubTimerTotalDisplay.data() + mFrameSubTimerTotalDisplay.length();
